implement tests for normalize function
