+++ src/test/java/de/fraunhofer/aisec/codyze/crymlin/BuildTranslationConfigTest.kt
@@ -88,6 +88,29 @@ internal class BuildTranslationConfigTest {
         )
     }
 
+
+    @Test
+    fun useAdditionalLanguagePython() {
+        val cliParameters = arrayOf("-c", "--additional-languages=python")
+        val config = Configuration.initConfig(null, *cliParameters)
+        config.buildServerConfiguration()
+
+        // able to handle missing frontends without crashing
+        assert(true)
+    }
+
+    @Test
+    fun useUnityBuildTest() {
+        val cliParameters = arrayOf("-c", "--unity", "--source=test.java")
+        val config = Configuration.initConfig(null, *cliParameters)
+        val translationConfig = config.buildTranslationConfiguration()
+
+        assertTrue(
+                translationConfig.loadIncludes,
+                "LoadIncludes has to be set to true if unityBuild is enabled"
+        )
+    }
+
     companion object {
         private lateinit var additionalOptionFile: File
